+++ pages/admin/master/voucher/index.jsx
@@ -13,6 +13,7 @@ import { AdminLayout } from '../../../../components/construct.components/layout/
 import { token_cookie_name } from '../../../../helpers';
 import { Decrypt } from '../../../../helpers/encryption.helpers';
 import { useRouter } from 'next/router';
+import VoucherForm from '../../../../components/voucher/VoucherForm';
 
 /* -------------------- Helpers -------------------- */
 
@@ -158,7 +159,7 @@ function VoucherCrud() {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [apiBase]);
 
-  // Upload image
+  // Upload image (hanya untuk local preview kalau kamu tetap pakai formData di halaman ini)
   const handleImageChange = (e) => {
     if (e.target.files && e.target.files[0]) {
       setImageFile(e.target.files[0]);
@@ -166,67 +167,29 @@ function VoucherCrud() {
     }
   };
 
-  // Submit (create / update)
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    if (formData.target_type === 'user' && !formData.target_user_id) {
-      alert('Pilih pengguna untuk target_type=user');
-      return;
-    }
-    if (formData.target_type === 'community' && !formData.community_id) {
-      alert('Pilih community untuk target_type=community');
-      return;
-    }
-
+  // Submit (create / update) — dipakai oleh VoucherForm via onSubmit
+  const submitVoucher = async (body) => {
     const isUpdate = Boolean(selectedVoucher);
     const url = isUpdate
       ? `${apiBase}/api/admin/vouchers/${selectedVoucher.id}`
       : `${apiBase}/api/admin/vouchers`;
 
-    const body = new FormData();
-    body.append('name', formData.name);
-    if (formData.description) body.append('description', formData.description);
-    if (formData.type) body.append('type', formData.type);
-    if (formData.valid_until) body.append('valid_until', formData.valid_until);
-    if (formData.tenant_location) body.append('tenant_location', formData.tenant_location);
-    body.append('stock', String(formData.stock ?? 0));
-    body.append('code', formData.code);
+    const res = await fetch(url, { method: 'POST', headers: { ...authHeader() }, body });
 
-    body.append('target_type', formData.target_type || 'all');
-    if (formData.target_type === 'user' && formData.target_user_id) {
-      body.append('target_user_id', formData.target_user_id);
-    }
-    if (formData.target_type === 'community' && formData.community_id) {
-      body.append('community_id', formData.community_id);
-    }
-
-    if (imageFile) body.append('image', imageFile);
-    if (isUpdate) body.append('_method', 'PUT');
-
-    try {
-      const res = await fetch(url, { method: 'POST', headers: { ...authHeader() }, body });
-      const text = await res.text();
-      let result;
+    if (!res.ok) {
+      let msg = 'server error';
       try {
-        result = text ? JSON.parse(text) : {};
-      } catch {
-        result = { raw: text };
-      }
-
-      if (!res.ok) {
-        const msg = result?.message || 'server error';
-        alert('Gagal menyimpan voucher: ' + msg);
-        return;
-      }
-
-      resetForm();
-      setRefreshToggle((s) => !s);
-      setModalForm(false);
-      fetchVouchers();
-    } catch {
-      alert('Terjadi kesalahan jaringan');
+        const j = await res.json();
+        msg = j?.message || msg;
+      } catch {}
+      alert('Gagal menyimpan voucher: ' + msg);
+      return;
     }
+
+    resetForm();
+    setRefreshToggle((s) => !s);
+    setModalForm(false);
+    fetchVouchers();
   };
 
   const resetForm = () => {
@@ -333,24 +296,11 @@ function VoucherCrud() {
         searchable
         noControlBar={false}
         setToRefresh={refreshToggle}
-        // Nonaktifkan "detail" & "add" bawaan, BIARKAN "edit" tampil lalu redirect ke page edit
         actionControl={{
           except: ['detail', 'add'],
           onEdit: (voucher) => {
             setSelectedVoucher(voucher);
-            setFormData({
-              name: voucher.name || '',
-              description: voucher.description || '',
-              image: voucher.image || '',
-              type: voucher.type || '',
-              valid_until: voucher.valid_until ? toDateInputValue(voucher.valid_until) : '',
-              tenant_location: voucher.tenant_location || '',
-              stock: voucher.stock ?? 0,
-              code: voucher.code || '',
-              community_id: voucher.community_id || '',
-              target_type: voucher.target_type || 'all',
-              target_user_id: voucher.target_user_id || '',
-            });
+            // formData di halaman ini tidak lagi menentukan UI; VoucherForm akan baca dari selectedVoucher via initialData
             setModalForm(true);
           },
           onDelete: (voucher) => {
@@ -367,270 +317,37 @@ function VoucherCrud() {
         }}
       />
 
-      {/* Modal Tambah Baru (create tetap di halaman ini) */}
+      {/* Modal Tambah / Edit — UI sama karena 1 komponen */}
       <FloatingPageComponent
         show={modalForm}
         onClose={() => {
           setModalForm(false);
           resetForm();
         }}
-        title={selectedVoucher ? 'Ubah Voucher' : 'Tambah Voucher'}
-        subtitle="Masukkan data yang valid dan benar!"
+        title="" // judul/badge dipindah ke VoucherForm agar konsisten
+        subtitle=""
         size="lg"
         className="bg-background"
       >
-        {/* Badge mode di pojok, sama kayak promo */}
-        <div className="px-6 pt-4">
-          <span className={`badge ${selectedVoucher ? 'badge-warning' : 'badge-success'}`}>
-            {selectedVoucher ? 'Mode Ubah' : 'Mode Tambah'}
-          </span>
-        </div>
-
-        <form className="p-6">
-          {/* Section: Informasi Umum */}
-          <div className="mb-4">
-            <h3 className="text-base font-semibold">Informasi Umum</h3>
-            <p className="text-sm text-muted-foreground">
-              Nama, deskripsi, dan identitas voucher.
-            </p>
-          </div>
-
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-            <div className="form-control">
-              <label className="label"><span className="label-text font-medium">Nama Voucher</span></label>
-              <input
-                type="text"
-                className="input input-bordered w-full"
-                placeholder="Contoh: Diskon 20% Semua Menu"
-                value={formData.name}
-                onChange={(e) => setFormData({ ...formData, name: e.target.value })}
-                required
-              />
-              <span className="text-xs text-gray-500 mt-1">Nama akan tampil di daftar voucher.</span>
-            </div>
-
-            <div className="form-control">
-              <label className="label"><span className="label-text font-medium">Kode Unik</span></label>
-              <input
-                type="text"
-                className="input input-bordered w-full"
-                placeholder="Contoh: HH-20OFF"
-                value={formData.code}
-                onChange={(e) => setFormData({ ...formData, code: e.target.value })}
-                required
-              />
-              <span className="text-xs text-gray-500 mt-1">Wajib & tidak boleh duplikat.</span>
-            </div>
-
-            <div className="md:col-span-2 form-control">
-              <label className="label"><span className="label-text font-medium">Deskripsi</span></label>
-              <input
-                type="text"
-                className="input input-bordered w-full"
-                placeholder="Tuliskan ketentuan singkat voucher"
-                value={formData.description}
-                onChange={(e) => setFormData({ ...formData, description: e.target.value })}
-              />
-            </div>
-          </div>
-
-          {/* Section: Pengaturan & Periode */}
-          <div className="mt-8 mb-4">
-            <h3 className="text-base font-semibold">Pengaturan & Periode</h3>
-            <p className="text-sm text-muted-foreground">Tipe, masa berlaku, lokasi, dan stok.</p>
-          </div>
-
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-            <div className="form-control">
-              <label className="label"><span className="label-text font-medium">Tipe Voucher</span></label>
-              <input
-                type="text"
-                className="input input-bordered w-full"
-                placeholder="Contoh: percentage / nominal / buy1get1"
-                value={formData.type}
-                onChange={(e) => setFormData({ ...formData, type: e.target.value })}
-              />
-            </div>
-
-            <div className="form-control">
-              <label className="label"><span className="label-text font-medium">Berlaku Sampai</span></label>
-              <input
-                type="date"
-                className="input input-bordered w-full"
-                value={formData.valid_until}
-                onChange={(e) => setFormData({ ...formData, valid_until: e.target.value })}
-              />
-              <span className="text-xs text-gray-500 mt-1">Format input YYYY-MM-DD.</span>
-            </div>
-
-            <div className="form-control">
-              <label className="label"><span className="label-text font-medium">Lokasi Tenant</span></label>
-              <input
-                type="text"
-                className="input input-bordered w-full"
-                placeholder="Contoh: Foodcourt Lantai 2"
-                value={formData.tenant_location}
-                onChange={(e) => setFormData({ ...formData, tenant_location: e.target.value })}
-              />
-            </div>
-
-            <div className="form-control">
-              <label className="label"><span className="label-text font-medium">Stok Voucher</span></label>
-              <div className="join w-full">
-                <input
-                  type="number"
-                  className="input input-bordered join-item w-full"
-                  min={0}
-                  value={formData.stock}
-                  onChange={(e) => setFormData({ ...formData, stock: Number(e.target.value) })}
-                  required
-                />
-                <span className="btn btn-ghost join-item pointer-events-none">voucher</span>
-              </div>
-              <span className="text-xs text-gray-500 mt-1">Tampilan tabel: “{`{angka}`} voucher”.</span>
-            </div>
-          </div>
-
-          {/* Section: Target Penerima */}
-          <div className="mt-8 mb-4">
-            <h3 className="text-base font-semibold">Target Penerima</h3>
-            <p className="text-sm text-muted-foreground">Atur siapa yang berhak menerima voucher.</p>
-          </div>
-
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-            <div className="form-control">
-              <label className="label"><span className="label-text font-medium">Target</span></label>
-              <select
-                className="select select-bordered w-full"
-                value={formData.target_type}
-                onChange={(e) => {
-                  const next = e.target.value;
-                  setFormData((s) => ({
-                    ...s,
-                    target_type: next,
-                    target_user_id: next === 'user' ? s.target_user_id : '',
-                    community_id: next === 'community' ? s.community_id : '',
-                  }));
-                }}
-              >
-                <option value="all">Semua Pengguna</option>
-                <option value="user">Pengguna Tertentu</option>
-                <option value="community">Anggota Community</option>
-              </select>
-            </div>
-
-            {formData.target_type === 'user' && (
-              <div className="form-control">
-                <label className="label"><span className="label-text font-medium">Pilih Pengguna</span></label>
-                <select
-                  className="select select-bordered w-full"
-                  value={formData.target_user_id}
-                  onChange={(e) => setFormData({ ...formData, target_user_id: e.target.value })}
-                  required
-                >
-                  <option value="">Pilih Pengguna</option>
-                  {users.map((u) => (
-                    <option key={u.id} value={u.id}>
-                      {u.name || u.email || `#${u.id}`}
-                    </option>
-                  ))}
-                </select>
-              </div>
-            )}
-
-            <div className="form-control md:col-span-2">
-              <label className="label">
-                <span className="label-text font-medium">
-                  Community {formData.target_type === 'community' ? '(wajib)' : '(opsional)'}
-                </span>
-              </label>
-              <select
-                className="select select-bordered w-full"
-                value={formData.community_id}
-                onChange={(e) => setFormData({ ...formData, community_id: e.target.value })}
-                required={formData.target_type === 'community'}
-                disabled={formData.target_type !== 'community'}
-              >
-                <option value="">Pilih Community</option>
-                {communities.map((c) => (
-                  <option key={c.id} value={c.id}>{c.name}</option>
-                ))}
-              </select>
-            </div>
-          </div>
-
-          {/* Section: Media */}
-          <div className="mt-8 mb-4">
-            <h3 className="text-base font-semibold">Media</h3>
-            <p className="text-sm text-muted-foreground">Gambar voucher (opsional).</p>
-          </div>
-
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-            <div className="form-control">
-              <label className="label"><span className="label-text font-medium">Upload Gambar</span></label>
-              <input
-                type="file"
-                accept="image/*"
-                className="file-input file-input-bordered w-full"
-                onChange={handleImageChange}
-              />
-              <span className="text-xs text-gray-500 mt-1">
-                PNG/JPG. Jika tidak diubah saat edit, gambar lama dipertahankan.
-              </span>
-            </div>
-
-            <div className="form-control">
-              <label className="label"><span className="label-text font-medium">Preview</span></label>
-              <div className="flex items-center gap-3">
-                {/* preview lama */}
-                {formData.image && !imageFile && (
-                  <img
-                    src={buildImageUrl(formData.image) || formData.image}
-                    alt="Voucher"
-                    width={96}
-                    height={96}
-                    className="rounded-xl border"
-                  />
-                )}
-                {/* preview baru */}
-                {imageFile && (
-                  <img
-                    src={URL.createObjectURL(imageFile)}
-                    alt="Preview"
-                    width={96}
-                    height={96}
-                    className="rounded-xl border"
-                  />
-                )}
-                {!formData.image && !imageFile && (
-                  <span className="text-sm text-muted-foreground">Belum ada gambar</span>
-                )}
-              </div>
-            </div>
-          </div>
-
-          {/* Footer Actions */}
-          <div className="mt-10 flex justify-end gap-2">
-            <ButtonComponent
-              label="Batal"
-              paint="secondary"
-              variant="outline"
-              onClick={() => {
-                setModalForm(false);
-                resetForm();
-              }}
-            />
-            <ButtonComponent
-              label={selectedVoucher ? 'Perbarui' : 'Simpan'}
-              paint="primary"
-              type="submit"
-              onClick={handleSubmit}
-            />
-          </div>
-        </form>
+        <VoucherForm
+          mode={selectedVoucher ? 'edit' : 'create'}
+          initialData={
+            selectedVoucher
+              ? { ...selectedVoucher, valid_until: selectedVoucher.valid_until ? selectedVoucher.valid_until : '' }
+              : undefined
+          }
+          users={users}
+          communities={communities}
+          buildImageUrl={buildImageUrl}
+          onCancel={() => {
+            setModalForm(false);
+            resetForm();
+          }}
+          onSubmit={submitVoucher}
+        />
       </FloatingPageComponent>
 
-      {/* Modal Delete Confirmation (pakai show + onSubmit) */}
+      {/* Modal Delete Confirmation */}
       <ModalConfirmComponent
         title="Hapus Voucher"
         show={modalDelete}
